+++ aw-reporting-model/src/main/java/com/google/api/ads/adwords/awreporting/model/entities/AgeRangePerformanceReport.java
@@ -18,7 +18,7 @@ import com.google.api.ads.adwords.awreporting.model.csv.annotation.CsvField;
 import com.google.api.ads.adwords.awreporting.model.csv.annotation.CsvReport;
 import com.google.api.ads.adwords.awreporting.model.csv.annotation.MoneyField;
 import com.google.api.ads.adwords.awreporting.model.util.BigDecimalUtil;
-import com.google.api.ads.adwords.lib.jaxb.v201710.ReportDefinitionReportType;
+import com.google.api.ads.adwords.lib.jaxb.v201802.ReportDefinitionReportType;
 
 import java.math.BigDecimal;
 
@@ -144,13 +144,21 @@ public class AgeRangePerformanceReport extends DateReport {
   @CsvField(value = "Base Campaign ID", reportField = "BaseCampaignId")
   private Long baseCampaignId;
 
+  @Column(name = "BiddingStrategyId")
+  @CsvField(value = "Bid Strategy ID", reportField = "BiddingStrategyId")
+  private Long biddingStrategyId;
+
+  @Column(name = "BiddingStrategyName")
+  @CsvField(value = "Bid Strategy Name", reportField = "BiddingStrategyName")
+  private String biddingStrategyName;
+
+  @Column(name = "BiddingStrategyType")
+  @CsvField(value = "Bid Strategy Type", reportField = "BiddingStrategyType")
+  private String biddingStrategyType;
+
   @Column(name = "BidModifier")
   @CsvField(value = "Bid adj.", reportField = "BidModifier")
   private BigDecimal bidModifier;
-
-  @Column(name = "BidType")
-  @CsvField(value = "Conversion optimizer bid type", reportField = "BidType")
-  private String bidType;
 
   @Column(name = "CampaignId")
   @CsvField(value = "Campaign ID", reportField = "CampaignId")
@@ -607,6 +615,30 @@ public class AgeRangePerformanceReport extends DateReport {
     this.baseCampaignId = baseCampaignId;
   }
 
+  public Long getBiddingStrategyId() {
+    return biddingStrategyId;
+  }
+
+  public void setBiddingStrategyId(Long biddingStrategyId) {
+    this.biddingStrategyId = biddingStrategyId;
+  }
+
+  public String getBiddingStrategyName() {
+    return biddingStrategyName;
+  }
+
+  public void setBiddingStrategyName(String biddingStrategyName) {
+    this.biddingStrategyName = biddingStrategyName;
+  }
+
+  public String getBiddingStrategyType() {
+    return biddingStrategyType;
+  }
+
+  public void setBiddingStrategyType(String biddingStrategyType) {
+    this.biddingStrategyType = biddingStrategyType;
+  }
+
   public String getBidModifier() {
     return BigDecimalUtil.formatAsReadable(bidModifier);
   }
@@ -617,14 +649,6 @@ public class AgeRangePerformanceReport extends DateReport {
 
   public void setBidModifier(String bidModifier) {
     this.bidModifier = BigDecimalUtil.parseFromNumberString(bidModifier);
-  }
-
-  public String getBidType() {
-    return bidType;
-  }
-
-  public void setBidType(String bidType) {
-    this.bidType = bidType;
   }
 
   public Long getCampaignId() {
@@ -1175,8 +1199,10 @@ public class AgeRangePerformanceReport extends DateReport {
       .append(averageCpv, other.averageCpv)
       .append(baseAdGroupId, other.baseAdGroupId)
       .append(baseCampaignId, other.baseCampaignId)
+      .append(biddingStrategyId, other.biddingStrategyId)
+      .append(biddingStrategyName, other.biddingStrategyName)
+      .append(biddingStrategyType, other.biddingStrategyType)
       .append(bidModifier, other.bidModifier)
-      .append(bidType, other.bidType)
       .append(campaignId, other.campaignId)
       .append(campaignName, other.campaignName)
       .append(campaignStatus, other.campaignStatus)
@@ -1261,8 +1287,10 @@ public class AgeRangePerformanceReport extends DateReport {
       .append(averageCpv)
       .append(baseAdGroupId)
       .append(baseCampaignId)
+      .append(biddingStrategyId)
+      .append(biddingStrategyName)
+      .append(biddingStrategyType)
       .append(bidModifier)
-      .append(bidType)
       .append(campaignId)
       .append(campaignName)
       .append(campaignStatus)