+++ aw-reporting-model/src/main/java/com/google/api/ads/adwords/awreporting/model/entities/CampaignNegativeKeywordsPerformanceReport.java
@@ -16,7 +16,7 @@ package com.google.api.ads.adwords.awreporting.model.entities;
 
 import com.google.api.ads.adwords.awreporting.model.csv.annotation.CsvField;
 import com.google.api.ads.adwords.awreporting.model.csv.annotation.CsvReport;
-import com.google.api.ads.adwords.lib.jaxb.v201710.ReportDefinitionReportType;
+import com.google.api.ads.adwords.lib.jaxb.v201802.ReportDefinitionReportType;
 
 
 import org.apache.commons.lang3.StringUtils;