+++ aw-reporting-model/src/main/java/com/google/api/ads/adwords/awreporting/model/entities/ParentalStatusPerformanceReport.java
@@ -18,7 +18,7 @@ import com.google.api.ads.adwords.awreporting.model.csv.annotation.CsvField;
 import com.google.api.ads.adwords.awreporting.model.csv.annotation.CsvReport;
 import com.google.api.ads.adwords.awreporting.model.csv.annotation.MoneyField;
 import com.google.api.ads.adwords.awreporting.model.util.BigDecimalUtil;
-import com.google.api.ads.adwords.lib.jaxb.v201710.ReportDefinitionReportType;
+import com.google.api.ads.adwords.lib.jaxb.v201802.ReportDefinitionReportType;
 
 import java.math.BigDecimal;
 
@@ -144,13 +144,21 @@ public class ParentalStatusPerformanceReport extends DateReport {
   @CsvField(value = "Base Campaign ID", reportField = "BaseCampaignId")
   private Long baseCampaignId;
 
+  @Column(name = "BiddingStrategyId")
+  @CsvField(value = "Bid Strategy ID", reportField = "BiddingStrategyId")
+  private Long biddingStrategyId;
+
+  @Column(name = "BiddingStrategyName")
+  @CsvField(value = "Bid Strategy Name", reportField = "BiddingStrategyName")
+  private String biddingStrategyName;
+
+  @Column(name = "BiddingStrategyType")
+  @CsvField(value = "Bid Strategy Type", reportField = "BiddingStrategyType")
+  private String biddingStrategyType;
+
   @Column(name = "BidModifier")
   @CsvField(value = "Bid adj.", reportField = "BidModifier")
   private BigDecimal bidModifier;
-
-  @Column(name = "BidType")
-  @CsvField(value = "Conversion optimizer bid type", reportField = "BidType")
-  private String bidType;
 
   @Column(name = "CampaignId")
   @CsvField(value = "Campaign ID", reportField = "CampaignId")
@@ -183,6 +191,10 @@ public class ParentalStatusPerformanceReport extends DateReport {
   @Column(name = "Conversions")
   @CsvField(value = "Conversions", reportField = "Conversions")
   private BigDecimal conversions;
+
+  @Column(name = "ConversionTrackerId")
+  @CsvField(value = "Conversion Tracker Id", reportField = "ConversionTrackerId")
+  private Long conversionTrackerId;
 
   @Column(name = "ConversionTypeName")
   @CsvField(value = "Conversion name", reportField = "ConversionTypeName")
@@ -603,6 +615,30 @@ public class ParentalStatusPerformanceReport extends DateReport {
     this.baseCampaignId = baseCampaignId;
   }
 
+  public Long getBiddingStrategyId() {
+    return biddingStrategyId;
+  }
+
+  public void setBiddingStrategyId(Long biddingStrategyId) {
+    this.biddingStrategyId = biddingStrategyId;
+  }
+
+  public String getBiddingStrategyName() {
+    return biddingStrategyName;
+  }
+
+  public void setBiddingStrategyName(String biddingStrategyName) {
+    this.biddingStrategyName = biddingStrategyName;
+  }
+
+  public String getBiddingStrategyType() {
+    return biddingStrategyType;
+  }
+
+  public void setBiddingStrategyType(String biddingStrategyType) {
+    this.biddingStrategyType = biddingStrategyType;
+  }
+
   public String getBidModifier() {
     return BigDecimalUtil.formatAsReadable(bidModifier);
   }
@@ -613,14 +649,6 @@ public class ParentalStatusPerformanceReport extends DateReport {
 
   public void setBidModifier(String bidModifier) {
     this.bidModifier = BigDecimalUtil.parseFromNumberString(bidModifier);
-  }
-
-  public String getBidType() {
-    return bidType;
-  }
-
-  public void setBidType(String bidType) {
-    this.bidType = bidType;
   }
 
   public Long getCampaignId() {
@@ -693,6 +721,14 @@ public class ParentalStatusPerformanceReport extends DateReport {
 
   public void setConversions(String conversions) {
     this.conversions = BigDecimalUtil.parseFromNumberString(conversions);
+  }
+
+  public Long getConversionTrackerId() {
+    return conversionTrackerId;
+  }
+
+  public void setConversionTrackerId(Long conversionTrackerId) {
+    this.conversionTrackerId = conversionTrackerId;
   }
 
   public String getConversionTypeName() {
@@ -1115,6 +1151,9 @@ public class ParentalStatusPerformanceReport extends DateReport {
     if (!StringUtils.isEmpty(conversionCategoryName)) {
       idBuilder.append("-").append(conversionCategoryName);
     }
+    if (conversionTrackerId != null) {
+      idBuilder.append("-").append(conversionTrackerId);
+    }
     if (!StringUtils.isEmpty(conversionTypeName)) {
       idBuilder.append("-").append(conversionTypeName);
     }
@@ -1160,8 +1199,10 @@ public class ParentalStatusPerformanceReport extends DateReport {
       .append(averageCpv, other.averageCpv)
       .append(baseAdGroupId, other.baseAdGroupId)
       .append(baseCampaignId, other.baseCampaignId)
+      .append(biddingStrategyId, other.biddingStrategyId)
+      .append(biddingStrategyName, other.biddingStrategyName)
+      .append(biddingStrategyType, other.biddingStrategyType)
       .append(bidModifier, other.bidModifier)
-      .append(bidType, other.bidType)
       .append(campaignId, other.campaignId)
       .append(campaignName, other.campaignName)
       .append(campaignStatus, other.campaignStatus)
@@ -1170,6 +1211,7 @@ public class ParentalStatusPerformanceReport extends DateReport {
       .append(conversionCategoryName, other.conversionCategoryName)
       .append(conversionRate, other.conversionRate)
       .append(conversions, other.conversions)
+      .append(conversionTrackerId, other.conversionTrackerId)
       .append(conversionTypeName, other.conversionTypeName)
       .append(conversionValue, other.conversionValue)
       .append(cost, other.cost)
@@ -1245,8 +1287,10 @@ public class ParentalStatusPerformanceReport extends DateReport {
       .append(averageCpv)
       .append(baseAdGroupId)
       .append(baseCampaignId)
+      .append(biddingStrategyId)
+      .append(biddingStrategyName)
+      .append(biddingStrategyType)
       .append(bidModifier)
-      .append(bidType)
       .append(campaignId)
       .append(campaignName)
       .append(campaignStatus)
@@ -1255,6 +1299,7 @@ public class ParentalStatusPerformanceReport extends DateReport {
       .append(conversionCategoryName)
       .append(conversionRate)
       .append(conversions)
+      .append(conversionTrackerId)
       .append(conversionTypeName)
       .append(conversionValue)
       .append(cost)