+++ aw-reporting-model/src/main/java/com/google/api/ads/adwords/awreporting/model/persistence/sql/SqlReportEntitiesPersister.java
@@ -20,22 +20,26 @@ import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
 import java.util.List;
 import org.hibernate.Criteria;
+import org.hibernate.NonUniqueObjectException;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.hibernate.criterion.Restrictions;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 /**
- * This is the basic implementation of the persistence layer to communicate with a SQL database.
- * The communication is done using a generic {@link SessionFactory}.
+ * This is the basic implementation of the persistence layer to communicate with a SQL database. The
+ * communication is done using a generic {@link SessionFactory}.
  */
 @Component
 @Qualifier("sqlEntitiesPersister")
 public class SqlReportEntitiesPersister implements EntityPersister {
 
+  private static final Logger logger = LoggerFactory.getLogger(SqlReportEntitiesPersister.class);
   private static final int BATCH_SIZE = 50;
 
   private int batchSize = BATCH_SIZE;
@@ -52,9 +56,7 @@ public class SqlReportEntitiesPersister implements EntityPersister {
         Preconditions.checkNotNull(sessionFactory, "SessionFactory can not be null");
   }
 
-  /**
-   * Persists all the given entities into the DB configured in the {@code SessionFactory}.
-   */
+  /** Persists all the given entities into the DB configured in the {@code SessionFactory}. */
   @Override
   @Transactional
   public void persistReportEntities(List<? extends Report> reportEntities) {
@@ -63,7 +65,31 @@ public class SqlReportEntitiesPersister implements EntityPersister {
 
     for (Report report : reportEntities) {
       report.setRowId();
-      session.saveOrUpdate(report);
+
+      try {
+        session.saveOrUpdate(report);
+      } catch (NonUniqueObjectException ex) {
+        // Github issue 268 & 280
+        //   https://github.com/googleads/aw-reporting/issues/268
+        //   https://github.com/googleads/aw-reporting/issues/280
+        //
+        // Currently we allow specifying report definitions which do not include all primary key
+        // fields. This leads to cryptic hibernate errors without providing a reasonable resolution
+        // strategy.
+        //
+        // This fix explains where to find the list of primary key fields, but does not address the
+        // underlying issue of allowing non-unique rows to be downloaded in the first place.
+        //
+        // Ideally we would guarantee uniqueness of rows without the user having to specify the
+        // PK fields. However, this would be a substantial migration for the AWReporting user base.
+        // Instead, we just log a (hopefully) useful error message.
+        logger.error(
+            "Duplicate row detected. This is most likely because your report definition does not "
+            + "include the primary key fields defined in {}.setRowId(). "
+            + "Please add the missing fields and try again.",
+            report.getClass().getName());
+        throw ex;
+      }
       batchFlush++;
 
       if (batchFlush == batchSize) {