+++ aw-reporting/src/test/java/com/google/api/ads/adwords/awreporting/util/ManagedCustomerServiceDelegateTest.java
@@ -19,11 +19,11 @@ import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import com.google.api.ads.adwords.jaxws.v201710.cm.Selector;
-import com.google.api.ads.adwords.jaxws.v201710.mcm.ApiException;
-import com.google.api.ads.adwords.jaxws.v201710.mcm.ManagedCustomer;
-import com.google.api.ads.adwords.jaxws.v201710.mcm.ManagedCustomerPage;
-import com.google.api.ads.adwords.jaxws.v201710.mcm.ManagedCustomerServiceInterface;
+import com.google.api.ads.adwords.jaxws.v201802.cm.Selector;
+import com.google.api.ads.adwords.jaxws.v201802.mcm.ApiException;
+import com.google.api.ads.adwords.jaxws.v201802.mcm.ManagedCustomer;
+import com.google.api.ads.adwords.jaxws.v201802.mcm.ManagedCustomerPage;
+import com.google.api.ads.adwords.jaxws.v201802.mcm.ManagedCustomerServiceInterface;
 import com.google.api.ads.adwords.lib.client.AdWordsSession;
 import java.util.Set;
 import org.junit.Before;